# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Aman's Watch Vault API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type MediaItemType = typeof MediaItemType[keyof typeof MediaItemType];


export const MediaItemType = {
  anime: 'anime',
  series: 'series',
} as const;

export type MediaItemStatus = typeof MediaItemStatus[keyof typeof MediaItemStatus];


export const MediaItemStatus = {
  completed: 'completed',
  watching: 'watching',
  plan_to_watch: 'plan_to_watch',
  dropped: 'dropped',
} as const;

export interface MediaItem {
  id: number;
  title: string;
  type: MediaItemType;
  year: number;
  /** @nullable */
  poster?: string | null;
  /** @nullable */
  banner?: string | null;
  /** @nullable */
  rating?: number | null;
  genres: string[];
  /** @nullable */
  synopsis?: string | null;
  /** @nullable */
  review?: string | null;
  isFavorite: boolean;
  /** @nullable */
  rank?: number | null;
  /** @nullable */
  episodes?: number | null;
  /** @nullable */
  seasons?: number | null;
  status: MediaItemStatus;
  createdAt: string;
  updatedAt: string;
}

export type MediaInputType = typeof MediaInputType[keyof typeof MediaInputType];


export const MediaInputType = {
  anime: 'anime',
  series: 'series',
} as const;

export type MediaInputStatus = typeof MediaInputStatus[keyof typeof MediaInputStatus];


export const MediaInputStatus = {
  completed: 'completed',
  watching: 'watching',
  plan_to_watch: 'plan_to_watch',
  dropped: 'dropped',
} as const;

export interface MediaInput {
  /** @minLength 1 */
  title: string;
  type: MediaInputType;
  year: number;
  poster?: string;
  banner?: string;
  /**
     * @minimum 0
     * @maximum 10
     */
  rating?: number;
  genres?: string[];
  synopsis?: string;
  review?: string;
  isFavorite?: boolean;
  rank?: number;
  episodes?: number;
  seasons?: number;
  status?: MediaInputStatus;
}

export type MediaUpdateType = typeof MediaUpdateType[keyof typeof MediaUpdateType];


export const MediaUpdateType = {
  anime: 'anime',
  series: 'series',
} as const;

export type MediaUpdateStatus = typeof MediaUpdateStatus[keyof typeof MediaUpdateStatus];


export const MediaUpdateStatus = {
  completed: 'completed',
  watching: 'watching',
  plan_to_watch: 'plan_to_watch',
  dropped: 'dropped',
} as const;

export interface MediaUpdate {
  /** @minLength 1 */
  title?: string;
  type?: MediaUpdateType;
  year?: number;
  poster?: string;
  banner?: string;
  /**
     * @minimum 0
     * @maximum 10
     */
  rating?: number;
  genres?: string[];
  synopsis?: string;
  review?: string;
  isFavorite?: boolean;
  rank?: number;
  episodes?: number;
  seasons?: number;
  status?: MediaUpdateStatus;
}

export interface RatingBucket {
  range: string;
  count: number;
}

export interface GenreCount {
  genre: string;
  count: number;
}

export interface VaultStats {
  totalAnime: number;
  totalSeries: number;
  totalItems: number;
  /** @nullable */
  averageRating?: number | null;
  highestRated?: MediaItem;
  totalCompleted: number;
  totalWatching: number;
  totalPlanToWatch: number;
  favoriteCount: number;
  ratingDistribution: RatingBucket[];
  genreBreakdown: GenreCount[];
}

export type ListMediaParams = {
type?: ListMediaType;
favorite?: boolean;
search?: string;
sort?: ListMediaSort;
order?: ListMediaOrder;
};

export type ListMediaType = typeof ListMediaType[keyof typeof ListMediaType];


export const ListMediaType = {
  anime: 'anime',
  series: 'series',
} as const;

export type ListMediaSort = typeof ListMediaSort[keyof typeof ListMediaSort];


export const ListMediaSort = {
  rating: 'rating',
  year: 'year',
  title: 'title',
  rank: 'rank',
} as const;

export type ListMediaOrder = typeof ListMediaOrder[keyof typeof ListMediaOrder];


export const ListMediaOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;

export type GetTopMediaParams = {
limit?: number;
};