# lib/api-zod/src/generated/types/mediaInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Aman's Watch Vault API
 * OpenAPI spec version: 0.1.0
 */
import type { MediaInputStatus } from './mediaInputStatus';
import type { MediaInputType } from './mediaInputType';

export interface MediaInput {
  /** @minLength 1 */
  title: string;
  type: MediaInputType;
  year: number;
  poster?: string;
  banner?: string;
  /**
     * @minimum 0
     * @maximum 10
     */
  rating?: number;
  genres?: string[];
  synopsis?: string;
  review?: string;
  isFavorite?: boolean;
  rank?: number;
  episodes?: number;
  seasons?: number;
  status?: MediaInputStatus;
}