# lib/api-zod/src/generated/types/mediaItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Aman's Watch Vault API
 * OpenAPI spec version: 0.1.0
 */
import type { MediaItemStatus } from './mediaItemStatus';
import type { MediaItemType } from './mediaItemType';

export interface MediaItem {
  id: number;
  title: string;
  type: MediaItemType;
  year: number;
  /** @nullable */
  poster?: string | null;
  /** @nullable */
  banner?: string | null;
  /** @nullable */
  rating?: number | null;
  genres: string[];
  /** @nullable */
  synopsis?: string | null;
  /** @nullable */
  review?: string | null;
  isFavorite: boolean;
  /** @nullable */
  rank?: number | null;
  /** @nullable */
  episodes?: number | null;
  /** @nullable */
  seasons?: number | null;
  status: MediaItemStatus;
  createdAt: Date;
  updatedAt: Date;
}