# lib/api-zod/src/generated/types/mediaUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Aman's Watch Vault API
 * OpenAPI spec version: 0.1.0
 */
import type { MediaUpdateStatus } from './mediaUpdateStatus';
import type { MediaUpdateType } from './mediaUpdateType';

export interface MediaUpdate {
  /** @minLength 1 */
  title?: string;
  type?: MediaUpdateType;
  year?: number;
  poster?: string;
  banner?: string;
  /**
     * @minimum 0
     * @maximum 10
     */
  rating?: number;
  genres?: string[];
  synopsis?: string;
  review?: string;
  isFavorite?: boolean;
  rank?: number;
  episodes?: number;
  seasons?: number;
  status?: MediaUpdateStatus;
}