# lib/api-zod/src/generated/types/ratingBucket.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Aman's Watch Vault API
 * OpenAPI spec version: 0.1.0
 */

export interface RatingBucket {
  range: string;
  count: number;
}