# lib/api-zod/src/generated/types/vaultStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Aman's Watch Vault API
 * OpenAPI spec version: 0.1.0
 */
import type { GenreCount } from './genreCount';
import type { MediaItem } from './mediaItem';
import type { RatingBucket } from './ratingBucket';

export interface VaultStats {
  totalAnime: number;
  totalSeries: number;
  totalItems: number;
  /** @nullable */
  averageRating?: number | null;
  highestRated?: MediaItem;
  totalCompleted: number;
  totalWatching: number;
  totalPlanToWatch: number;
  favoriteCount: number;
  ratingDistribution: RatingBucket[];
  genreBreakdown: GenreCount[];
}